requirements.txt
# Core FastAPI framework and ASGI server
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# Data validation and serialization
pydantic>=2.5.0,<3.0.0

# Google Cloud AI Platform integration
google-cloud-aiplatform>=1.38.0,<2.0.0

# LangChain with Google Vertex AI support
langchain-google-vertexai>=0.0.6,<1.0.0
langchain>=0.1.0,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-openai>=0.0.5,<1.0.0
langchainhub>=0.1.14,<1.0.0

# Environment configuration
python-dotenv>=1.0.0,<2.0.0

# Additional dependencies for RAG and multi-agent systems
chromadb>=0.4.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0
numpy>=1.24.0,<2.0.0
tiktoken>=0.5.0,<1.0.0

# HTTP client for external API calls
httpx>=0.25.0,<1.0.0

# Logging and monitoring
structlog>=23.1.0,<24.0.0

# Security and rate limiting
slowapi>=0.1.9,<1.0.0

# File processing for knowledge base
python-multipart>=0.0.6,<1.0.0
pypdf>=3.17.0,<4.0.0
unstructured>=0.10.0,<1.0.0

# Document processing for RAG
langchain-community>=0.0.10,<1.0.0

# Telegram Bot API
python-telegram-bot>=20.0,<21.0.0